[fix] use long for id

+++ src/main/java/com/vinninho/demo/DemoController.java
@@ -21,7 +21,7 @@ public class DemoController {
     }
 
     @GetMapping("/find/{id}")
-    public Customer findCustomer(@PathVariable Integer id) {
+    public Customer findCustomer(@PathVariable Long id) {
         return customerRepository.findCustomerById(id);
     }
 }
